# data/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Provides all projects the opportunity to access Ethereum blockchain through RESTful API.",
    "title": "HOQU blockchain API by GoLang.",
    "contact": {
      "name": "Denis Degterin",
      "url": "https://github.com/hoqu-io/hoqu-geth-api",
      "email": "d.degterin@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.3.1"
  },
  "basePath": "/",
  "paths": {
...
    "/platform/user/:id": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Get User by ID.",
        "operationId": "getUser",
        "parameters": [
          {
            "type": "string",
            "example": "a6fdb91a-149d-11e8-b642-0ed5f89f718b",
            "x-go-name": "Id",
            "description": "an ID of the requested entity",
            "name": "id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserDataResponse"
          },
          "400": {
            "$ref": "#/responses/RestErrorResponse"
          }
        }
      }
    },
    "/platform/user/address": {
      "post": {
        "description": "Each User can have several Ethereum addresses.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Add User Ethereum Address.",
        "operationId": "addUserAddress",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddUserAddressRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TxSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/RestErrorResponse"
          }
        }
      }
    },
    "/platform/user/kyc": {
      "post": {
        "description": "Each User has KYC level which reflects his trust level.\nUser KYC level can be changed only by KYC reports.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Add User KYC report.",
        "operationId": "addUserKyc",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddUserKycReportRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TxSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/RestErrorResponse"
          }
        }
      }
    },
    "/platform/user/register": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Register User.",
        "operationId": "registerUser",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RegisterUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AddSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/RestErrorResponse"
          }
        }
      }
    },
    "/platform/user/status": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Set User Status.",
        "operationId": "setUserStatus",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SetStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TxSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/RestErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "AddUserAddressRequest": {
      "type": "object",
      "properties": {
        "address": {
          "description": "Ethereum address of the user",
          "type": "string",
          "x-go-name": "Address",
          "example": "0xED2F74E1fb73b775E6e35720869Ae7A7f4D755aD"
        },
        "id": {
          "description": "an ID of the user",
          "type": "string",
          "x-go-name": "Id",
          "example": "a6fdb91a-149d-11e8-b642-0ed5f89f718b"
        }
      },
      "x-go-package": "hoqu-geth-api/geth/models"
    },
    "AddUserKycReportRequest": {
      "type": "object",
      "properties": {
        "dataUrl": {
          "description": "the URL with full report data",
          "type": "string",
          "x-go-name": "DataUrl"
        },
        "id": {
          "description": "an ID of the user",
          "type": "string",
          "x-go-name": "Id",
          "example": "a6fdb91a-149d-11e8-b642-0ed5f89f718b"
        },
        "kycLevel": {
          "description": "Higher the level more trust this user has from other participants",
          "type": "integer",
          "format": "uint8",
          "x-go-name": "KycLevel",
          "example": 3
        },
        "meta": {
          "description": "Meta data of the report in string format (JSON for example)",
          "type": "string",
          "x-go-name": "Meta"
        }
      },
      "x-go-package": "hoqu-geth-api/geth/models"
    },
...
    "RegisterUserRequest": {
      "type": "object",
      "properties": {
        "address": {
          "description": "Ethereum address of the user",
          "type": "string",
          "x-go-name": "Address",
          "example": "0xED2F74E1fb73b775E6e35720869Ae7A7f4D755aD"
        },
        "pubKey": {
          "description": "all private data available only for that user is encrypted by his public key",
          "type": "string",
          "x-go-name": "PubKey"
        },
        "role": {
          "description": "user role",
          "type": "string",
          "x-go-name": "Role",
          "example": "affiliate"
        }
      },
      "x-go-package": "hoqu-geth-api/geth/models"
    },
    "SetStatusRequest": {
      "type": "object",
      "properties": {
        "id": {
          "description": "id of entity",
          "type": "string",
          "x-go-name": "Id",
          "example": "a6fdb91a-149d-11e8-b642-0ed5f89f718b"
        },
        "status": {
          "$ref": "#/definitions/Status"
        }
      },
      "x-go-package": "hoqu-geth-api/geth/models"
    },
    "Status": {
      "description": "status, one of 0:NotExists, 1:Created, 2:Pending, 3:Active, 4:Done, 5:Declined",
      "type": "integer",
      "format": "uint8",
      "x-go-package": "hoqu-geth-api/geth/models"
    },
...
    "UserData": {
      "description": "User Model",
      "type": "object",
      "properties": {
        "addresses": {
          "description": "a list of Ethereum addresses of the user",
          "x-go-name": "Addresses",
          "example": "[\"0xED2F74E1fb73b775E6e35720869Ae7A7f4D755aD\"]"
        },
        "createdAt": {
          "description": "unix timestamp",
          "type": "string",
          "x-go-name": "CreatedAt",
          "example": "1518957879"
        },
        "kycLevel": {
          "description": "Higher the level more trust this user has from other participants",
          "type": "integer",
          "format": "uint8",
          "x-go-name": "KycLevel",
          "example": 3
        },
        "kycReports": {
          "description": "a list of all KYC reports",
          "x-go-name": "KycReports"
        },
        "pubKey": {
          "description": "all private data available only for that user is encrypted by his public key",
          "type": "string",
          "x-go-name": "PubKey"
        },
        "role": {
          "description": "user role",
          "type": "string",
          "x-go-name": "Role",
          "example": "affiliate"
        },
        "status": {
          "$ref": "#/definitions/Status"
        }
      },
      "x-go-package": "hoqu-geth-api/geth/models"
    }
  },
  "responses": {
...
    "AddSuccessResponse": {
      "description": "Success",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "description": "id of created entity",
                "type": "string",
                "x-go-name": "Id",
                "example": "a6fdb91a-149d-11e8-b642-0ed5f89f718b"
              },
              "tx": {
                "description": "Ethereum transaction hash",
                "type": "string",
                "x-go-name": "Tx",
                "example": "0x23682ef776bfb465433e8f6a6e84eab71f039f039e86933aeca20ee46d01d576"
              }
            },
            "x-go-name": "Data"
          }
        }
      }
    },
...
    "RestErrorResponse": {
      "description": "Error",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "description": "error code",
                "type": "string",
                "x-go-name": "Code",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "description": "error description",
                "type": "string",
                "x-go-name": "Message",
                "example": "id should be in UUID format"
              }
            },
            "x-go-name": "Error"
          }
        }
      }
    },
...
    "TxSuccessResponse": {
      "description": "Success",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "tx": {
                "description": "Ethereum transaction hash",
                "type": "string",
                "x-go-name": "Tx",
                "example": "0x23682ef776bfb465433e8f6a6e84eab71f039f039e86933aeca20ee46d01d576"
              }
            },
            "x-go-name": "Data"
          }
        }
      }
    },
    "UserDataResponse": {
      "description": "Success",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "User": {
                "$ref": "#/definitions/UserData"
              }
            },
            "x-go-name": "Data"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "hash": {
      "type": "apiKey",
      "name": "X-Sign",
      "in": "header"
    },
    "jwt": {
      "type": "apiKey",
      "name": "X-Authorization",
      "in": "header"
    }
  }
}